+++ flask_unchained/bundles/controller/metaclasses.py
@@ -23,7 +23,7 @@ class ControllerMeta(type):
     - if subclass of a base class, init CONTROLLER_ROUTES_ATTR
         - check if methods were decorated with @route, otherwise
           create a new Route for each method
-        - finish initializing Routes (set blueprint, _controller_name)
+        - finish initializing Routes
     """
     def __new__(mcs, name, bases, clsdict):
         set_up_class_dependency_injection(name, clsdict)
@@ -44,14 +44,17 @@ class ControllerMeta(type):
                 controller_routes.pop(method_name, None)
                 continue
 
-            method_routes = getattr(method, FN_ROUTES_ATTR,
-                                    [Route(None, method)])
-            for route in method_routes:
-                route._controller_name = name
+            method_routes = getattr(method, FN_ROUTES_ATTR, [Route(None, method)])
             controller_routes[method_name] = method_routes
 
         setattr(cls, CONTROLLER_ROUTES_ATTR, controller_routes)
         return cls
+
+    def __init__(cls, name, bases, clsdict):
+        super().__init__(name, bases, clsdict)
+        for routes in getattr(cls, CONTROLLER_ROUTES_ATTR, {}).values():
+            for route in routes:
+                route._controller_cls = cls
 
 
 class ResourceMeta(ControllerMeta):